daemon: Fix |DPRINTF()|/|eprintf()| format type mismatches
Fix |DPRINTF()|/|eprintf()| format type mismatches, and use
Visual Sudio-specific macros to enable warnings in Analyzer
mode to catch any new mismatches.

Reported-by: Cedric Blancher <cedric.blancher@gmail.com>

+++ daemon/fsctl.c
@@ -458,7 +458,7 @@ static int parse_duplicatedata(unsigned char *buffer,
         "duplicatedata=(src_state=0x%p srcfileoffset=%lld "
         "destfileoffset=%lld bytecount=%lld)\n",
         opcode2string(upcall->opcode),
-        (long long)args->src_state,
+        args->src_state,
         (long long)args->srcfileoffset,
         (long long)args->destfileoffset,
         (long long)args->bytecount));
